Refactor product conversion and grossWeight to double
Changed the types of 'conversion' and 'grossWeight' from int to double in ProductItem and updated related widget logic to handle double values. This improves precision for product measurements and ensures consistent formatting and calculations in both desktop and mobile product card views.

+++ lib/features/products/data/models/product_item.dart
@@ -6,7 +6,7 @@ class ProductItem {
   int? packingTypeID;
   String? sellingUM;
   int? conversion;
-  int? grossWeight;
+  double? grossWeight;
   int? itemCategory1;
   int? itemCategory2;
   String? base;
@@ -119,7 +119,7 @@ class ProductItem {
     int? packingTypeID,
     String? sellingUM,
     int? conversion,
-    int? grossWeight,
+    double? grossWeight,
     int? itemCategory1,
     int? itemCategory2,
     String? base,
